requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
python-jose
passlib[bcrypt]
mysql-connector-python